BNX2X: IRQs implemented
IRQs are registered okay, but we were hanging here:

[bnx2x]:[bnx2x_state_wait:285()]waiting for state to become 1
[bnx2x]:[bnx2x_sp_task:5648()]sp task invoked
[bnx2x]:[bnx2x_sp_task:5657()]status 0
[bnx2x]:[bnx2x_sp_task:5658()]setting interrupt_occurred to 0
[bnx2x]:[bnx2x_state_wait:303()]timeout waiting for state 1

The issue is that we should be receiving an FP IRQ at this point, and we
needed to be consistent with whether or not we have CNIC_SUPPORT.  If we
do, then the code needs to reserve an extra MSIX vector in the
pci_devices table for the CNIC.  The actual card will know about it
(since we tell it elsewhere), and then the FP's won't begin until MSIX
vector 2 (the third one).

Also, since we do various hackery, we need to make sure the driver knows
we're using MSIX.  We'll tell it later, based on the flag.

Btw, the current pr_fmt seems a bit obnoxious.  Damned if you do, damned if you
don't.