BNX2X: limit queues to 2
Linux does not need to do this.  Something we're doing is wrong still.  If we
run with 8 queues, we get the dreaded MC assert.  With 4 total, we're okay (for
now).

For now, we're limiting to four queues.  I tried looking for other diffs
between Linux and Akaros with 8 queues, but haven't found the problem yet.

But that total includes both functions, so we have only two per controller.