Changes devether's Max Trans Unit to 1500
From 1514.  Linux drivers track it as 1500, and add on the header as
needed.  Plan 9 uses 1514, and includes the header.

I don't particularly care one way or another.  This way is a little
easier for porting, but have no problem with changing it back (given a
port to bnx2x's MTU code).