Fixes devnix's V lookup
The issue was that the QID for chans that were genned from a Qnixdir had
a bad nixid.  We were passing either 0 (for ctl) or 1 (for image) (which
is what you get when you take (s - Qctl).  This happened to work okay
for ctl, since we usually have a nix0.  But image thought its struct v
was for nix1.

I think we only need to do this for "state machine genned" entries,
meaning the entries under Qnixdir.  The "directly genned" entries should
come in to gen with a properly filled out QID (which came from a
previous gen).