Run a NIX guest.
This works. Basically, when a NIX is cloned, we put a simple program in it to test.

When that NIX is started, it puts a
0
on 3f8, and return.

We've seen this work.

In this case, I did this:
cat '#t/clone'
echo reserve 5 > '#t/nix0/ctl'
echo run 5 0xffff800040000000 > '#t/nix0/ctl'

Signed-off-by: Ron Minnich <rminnich@gmail.com>