Try out Dune vmm stuff.
This has been frustrating, finding a reasonably compact yet not awful bit of code
to use for akaros. They're either gigantic, featureful with stuff we don't want,
buggy as can be, or all three.

The Dune stuff seems a reasonably pared down subset of kvm and it's actually
pretty clean. Also, unlike every other bit of vmx code we've tried lately,
it actually seems to work, which is a bonus. And it will never run as a 1978 8086.

We actually start a VM at this point. And, it fails, and exits, *but akaros is not hurt*.
We just keep going.

This was imported, heavily edited, and retroactively committed.
(git-fu'd by brho)