Makes pte_t an opaque type
The nature of pte_t will be arch-dependent.  It could be a struct or the
old pte_t (kpte_t) and have its own tricks for modification.  This patch
changes all of the accesses that implicitly treated pte_t as something
that could be read or written in a certain manner.

It also catches a bug in the unused (and warned) unmap_vmap_segment().