x86: changes pte_t to be a KPTE and an EPTE
Ideally, a PTE would always have a KPTE, and optionally have an EPTE.
Now, for all ops other than walk_okay, we have at least one of KPTE or
EPTE (and possibly both).  pgdir_walk() will return both.  memwalks will
have only one or the other.

The alternative is to come up with a way to walk both tables in
lockstep, jumping back and forth and building a full PTE for the
memcallback at each point.  We might do that later, if necessary.