BNX2X: fixes cont_page allocation
Linux's alloc_pages returns a struct page*.  Our get_cont_pages returns
a KVA.  The mapping would be screwed up, since we were taking the
page2pa() of the KVA, not the page.

Of course, we weren't calling this, since TPA was disabled, and it
wasn't the problem triggering the MC Assert.  Well damn.