BNX2X: hacks around the MC assert problem
Our get_cont_pages() does not return pages aligned to the *order* of the
allocation.  This means higher order allocs, like 2^3 pages might not
come on an 8*PGSIZE boundary.  Apparently, this was enough to trigger
the MC assert for some high-order alloc.

The dirtiness comes in where we don't free the contig region.  Hopefully
we're not leaking too much.  (only once per NIC, so far).

This is mostly a stopgap til we fix our shitty memory allocator.