VMM: removes the VPID
AFAIK, the VPID is only useful if you aren't using EPT.  Since we're
using the EPT, the VPID code is just a waste.  But I did leave in the
helpers, since they may be useful and it's also a good practice to flush
the VPID caches in case a previous user of VMX messed with them.