VMM: Removes the epte_t from pte_t
A pte is just a kpte_t* - still different than a month ago, where pte ==
kpte.  Now pte == kpte*.  But we don't need to track the epte with the
kpte, since we can find the epte from the kpte.

I kept the eptp as part of the pgdir.  Arguably, the cr3 belongs there
too.  Often the cr3 and eptp have extra bits in them, other than just
the PADDR of the PML4, so it's handy to keep them around.