Builds libstdc++.so (XCC)
Changed up the configure for libstdc++ to match Linux's config.  If
you're porting the configure script to a newer version of gcc,
double-check all of the places where 'linux' appears in the config.
It's probably easier to just recreate the configure script from scratch.

There are a few places in the configure script where it's probably not
important for Akaros to mirror Linux, such as when testing for certain
Linux features.  Since those tests just fail, it seems saner just to put
an 'akaros' wherever there is a 'linux.'

Also, I had to update fill-kfs to find the libstdc++.so, and this also
adds about 6MB to KFS (and the resulting kernel binary) due to the size
of libstdc++.so.

Rebuild your toolchain.