Untangling the glibc header rat's nest (P.1) (XCC)
The problem we have is that there are various kernel headers and various
header files in glibc that cover the same thing, and sometimes they
differ.

For instance, there are the O_ flags for open.  Glibc had them in fcntl.
Akaros had them in fs.h.  It'd be nice if we used the same files and
structures!  This came up as a problem when the files start to diverge.
For instance, O_NOFOLLOW was in fs.h, but not in bits/fcntl.h.

Unfortunately, the headers are a rat's nest.  (Rats' nest?)  Making the
simple changes caused time.h to get pulled in early, and there's also a
macro for st_atime and friends (to st_atim.tv_sec).  Good times.  In the
end, handling time.h seems to work, but with a rename of st_atim{e,}.

bits/stat.h has similar issues.  In this commit, I took care of the
fcntl stuff, hence the P.1 (stands for Part 1).  One of these days we
can try to dig a bit further with other headers.  At least we're a
little better off now.

Rebuild your toolchain.