Moves KiB and friends from common.h (XCC)
These #defines mess with any app that tries to use those names for
something else.  And the sizes are not part of the kernel interface.

Arguably, we could get rid of them too, but we use a couple of them in a
few places.

You could rebuild your toolchain, or just copy the modified header file
(ros/common.h) into your toolchain installation sysroot.