User interface for fsync and posix_fadvise (XCC)
Using fcntl() for fsync and fadvise.  Not sure why POSIX/glibc didn't do
this already.

Part of it might be that fcntl, despite the ... declaration, supposedly
always takes one va_arg of type int, and it's ignored for certain
commands (e.g. F_GETFD).  That's pretty silly, so I extended it to take
up to four args.

Rebuild glibc.