vmmcp: this should be the complete list of registers we need to autoload
If we set FS_BASE and GS_BASE on autoload, vmmcpkernel causes
a vmware machine check. Why?

For LSTAR:
You have to disable intercept but also for this one do autoload.

If you don't disable intercept then the guest can not read or set it.
If you don't set autoload and the guest sets it, then this non-virtualized
register will probably kill your kernel.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>