Cleans up and renames ndblib
Ndblib was named ndb when it was installed; this changes it to ndblib.

We could either rename it completely to ndb (directory and everything)
and do the same for iplib, or just keep it called ndblib.  Given that
the libs are dropped into the sysroot directly, keeping them as iplib
and ndblib decreases the chance for collisions.

Moves ndbhf.h to an include/.  Any user .h file other than the pthread
headers needs to be in an include directory.