Fixes SYS_vc_entry numbering (XCC)
SYS_vc_entry was using the same number as SYS_abort_sysc.  I think I was
trying to keep vc_entry with similar calls, but didn't want to mess with
the numbers.  Clearly that was wrong, since I later used 31 again.

Incidentally, the last initializer wins for gcc, though a warning might
have been nice (not sure if there is a legitimate usage of this).
https://gcc.gnu.org/onlinedocs/gcc/Designated-Inits.html

Reinstall the kernel header and rebuild userspace.  No need to rebuild the
toolchain.