vmmcp: final set of changes for PCI emulation
In the end, this is not going to work. PCI in Linux brings in all sorts of stuff you don't want to bring in, such as ACPI. If
you don't bring in ACPI, it brings in dmi decode. Either way, it brings in code that wants to grovel through the f segment and
we've declared the low 16m off limits.

So, I'm going to take a different tack. We've got the coreboot tables and vmcall, might as well use them for something.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>
Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>